Refactor split Route and Routing classes
