+++ src/Router/Route.php
@@ -8,17 +8,6 @@ use Gacela\Resolver\InstanceCreator;
 
 use function is_object;
 
-/**
- * @method static get(string $path, object|string $controller, string $action = '__invoke')
- * @method static head(string $path, object|string $controller, string $action = '__invoke')
- * @method static connect(string $path, object|string $controller, string $action = '__invoke')
- * @method static post(string $path, object|string $controller, string $action = '__invoke')
- * @method static delete(string $path, object|string $controller, string $action = '__invoke')
- * @method static options(string $path, object|string $controller, string $action = '__invoke')
- * @method static patch(string $path, object|string $controller, string $action = '__invoke')
- * @method static put(string $path, object|string $controller, string $action = '__invoke')
- * @method static trace(string $path, object|string $controller, string $action = '__invoke')
- */
 final class Route
 {
     /**
@@ -30,21 +19,6 @@ final class Route
         private object|string $controller,
         private string $action = '__invoke',
     ) {
-    }
-
-    /**
-     * @param callable(RoutingConfigurator):void $fn
-     */
-    public static function configure(callable $fn): void
-    {
-        $routingConfigurator = new RoutingConfigurator();
-        $fn($routingConfigurator);
-
-        $route = self::findRoute($routingConfigurator);
-
-        if ($route) {
-            echo $route->run($routingConfigurator);
-        }
     }
 
     /**
@@ -89,33 +63,7 @@ final class Route
         return '#^/' . $pattern . '$#';
     }
 
-    private static function findRoute(RoutingConfigurator $routingConfigurator): ?self
-    {
-        foreach ($routingConfigurator->routes() as $route) {
-            $redirect = $routingConfigurator->redirects()[$route->path()] ?? null;
-            if ($redirect !== null) {
-                return self::findRedirectRoute($redirect, $routingConfigurator);
-            }
-
-            if ($route->requestMatches()) {
-                return $route;
-            }
-        }
-
-        return null;
-    }
-
-    private static function findRedirectRoute(Redirect $redirect, RoutingConfigurator $routingConfigurator): ?self
-    {
-        foreach ($routingConfigurator->routes() as $route) {
-            if ($route->path() === $redirect->destination()) {
-                return $route;
-            }
-        }
-        return null;
-    }
-
-    private function requestMatches(): bool
+    public function requestMatches(): bool
     {
         if (!$this->methodMatches()) {
             return false;
