+++ tests/Unit/Router/RoutingTest.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace GacelaTest\Unit\Router;
 
 use Gacela\Router\Request;
-use Gacela\Router\Route;
+use Gacela\Router\Routing;
 use Gacela\Router\RoutingConfigurator;
 use Gacela\Router\UnsupportedHttpMethodException;
 use GacelaTest\Unit\Router\Fake\Name;
@@ -13,7 +13,7 @@ use GacelaTest\Unit\Router\Fake\NameInterface;
 use Generator;
 use PHPUnit\Framework\TestCase;
 
-final class RouteTest extends TestCase
+final class RoutingTest extends TestCase
 {
     private const PROVIDER_TRIES = 10;
 
@@ -29,7 +29,7 @@ final class RouteTest extends TestCase
 
         $this->expectOutputString('Expected!');
 
-        Route::configure(static function (RoutingConfigurator $routes): void {
+        Routing::configure(static function (RoutingConfigurator $routes): void {
             $routes->get('expected/uri', FakeController::class, 'basicAction');
         });
     }
@@ -41,7 +41,7 @@ final class RouteTest extends TestCase
 
         $this->expectOutputString('');
 
-        Route::configure(static function (RoutingConfigurator $routes): void {
+        Routing::configure(static function (RoutingConfigurator $routes): void {
             $routes->get('other/uri', FakeController::class, 'basicAction');
         });
     }
@@ -53,7 +53,7 @@ final class RouteTest extends TestCase
 
         $this->expectOutputString('');
 
-        Route::configure(static function (RoutingConfigurator $routes): void {
+        Routing::configure(static function (RoutingConfigurator $routes): void {
             $routes->post('expected/uri', FakeController::class, 'basicAction');
         });
     }
@@ -68,7 +68,7 @@ final class RouteTest extends TestCase
 
         $this->expectOutputString("The params are '{$params[0]}', '{$params[1]}' and '{$params[2]}'!");
 
-        Route::configure(static function (RoutingConfigurator $routes): void {
+        Routing::configure(static function (RoutingConfigurator $routes): void {
             $routes->get('{firstParam}/{secondParam}/{thirdParam}', FakeController::class, 'manyParamsAction');
         });
     }
@@ -83,7 +83,7 @@ final class RouteTest extends TestCase
 
         $this->expectOutputString("The params are '{$params[1]}', '{$params[0]}' and '{$params[2]}'!");
 
-        Route::configure(static function (RoutingConfigurator $routes): void {
+        Routing::configure(static function (RoutingConfigurator $routes): void {
             $routes->get('{secondParam}/{firstParam}/{thirdParam}', FakeController::class, 'manyParamsAction');
         });
     }
@@ -98,7 +98,7 @@ final class RouteTest extends TestCase
 
         $this->expectOutputString("The 'string' param is '{$string}'!");
 
-        Route::configure(static function (RoutingConfigurator $routes): void {
+        Routing::configure(static function (RoutingConfigurator $routes): void {
             $routes->get('expected/string/is/{param}', FakeController::class, 'stringParamAction');
         });
     }
@@ -121,7 +121,7 @@ final class RouteTest extends TestCase
 
         $this->expectOutputString("The 'int' param is '{$int}'!");
 
-        Route::configure(static function (RoutingConfigurator $routes): void {
+        Routing::configure(static function (RoutingConfigurator $routes): void {
             $routes->get('expected/integer/is/{param}', FakeController::class, 'intParamAction');
         });
     }
@@ -144,7 +144,7 @@ final class RouteTest extends TestCase
 
         $this->expectOutputString("The 'float' param is '{$float}'!");
 
-        Route::configure(static function (RoutingConfigurator $routes): void {
+        Routing::configure(static function (RoutingConfigurator $routes): void {
             $routes->get('expected/float/is/{param}', FakeController::class, 'floatParamAction');
         });
     }
@@ -167,7 +167,7 @@ final class RouteTest extends TestCase
 
         $this->expectOutputString("The 'bool' param is '{$expected}'!");
 
-        Route::configure(static function (RoutingConfigurator $routes): void {
+        Routing::configure(static function (RoutingConfigurator $routes): void {
             $routes->get('expected/bool/is/{param}', FakeController::class, 'boolParamAction');
         });
     }
@@ -187,7 +187,7 @@ final class RouteTest extends TestCase
 
         $this->expectOutputString('Expected!');
 
-        Route::configure(static function (RoutingConfigurator $routes): void {
+        Routing::configure(static function (RoutingConfigurator $routes): void {
             $routes->get('expected/uri', FakeController::class, 'basicAction');
             $routes->get('expected/{param}', FakeController::class, 'stringParamAction');
         });
@@ -200,7 +200,7 @@ final class RouteTest extends TestCase
 
         $this->expectOutputString('Expected!');
 
-        Route::configure(static function (RoutingConfigurator $routes): void {
+        Routing::configure(static function (RoutingConfigurator $routes): void {
             $routes->get('optional/{param?}', FakeController::class, 'basicAction');
         });
     }
@@ -212,7 +212,7 @@ final class RouteTest extends TestCase
 
         $this->expectOutputString('Expected!');
 
-        Route::configure(static function (RoutingConfigurator $routes): void {
+        Routing::configure(static function (RoutingConfigurator $routes): void {
             $routes->get('optional/{param1?}/{param2?}', FakeController::class, 'basicAction');
         });
     }
@@ -221,7 +221,7 @@ final class RouteTest extends TestCase
     {
         $this->expectException(UnsupportedHttpMethodException::class);
 
-        Route::configure(static function (RoutingConfigurator $routes): void {
+        Routing::configure(static function (RoutingConfigurator $routes): void {
             $routes->invalidName('', FakeController::class);
         });
     }
@@ -236,7 +236,7 @@ final class RouteTest extends TestCase
 
         $this->expectOutputString('Expected!');
 
-        Route::configure(static function (RoutingConfigurator $routes): void {
+        Routing::configure(static function (RoutingConfigurator $routes): void {
             $routes->any('expected/uri', FakeController::class, 'basicAction');
         });
     }
@@ -261,7 +261,7 @@ final class RouteTest extends TestCase
 
         $this->expectOutputString('default-Expected!');
 
-        Route::configure(static function (RoutingConfigurator $routes): void {
+        Routing::configure(static function (RoutingConfigurator $routes): void {
             $routes->get('expected/uri', FakeControllerWithDependencies::class);
             $routes->setMappingInterfaces([NameInterface::class => new Name('Expected!')]);
         });
@@ -274,7 +274,7 @@ final class RouteTest extends TestCase
 
         $this->expectOutputString('Expected!');
 
-        Route::configure(static function (RoutingConfigurator $routes): void {
+        Routing::configure(static function (RoutingConfigurator $routes): void {
             $routes->redirect('optional/uri', 'expected/uri');
 
             $routes->get('expected/uri', FakeController::class, 'basicAction');
